+++ lib/lwm2m-node-lib.js
@@ -26,8 +26,8 @@
 exports.server = require('./lwm2m-server');
 exports.client = require('./lwm2m-client');
 
-var object = require('./services/object');
-exports.Schema = object.Schema; // constructor
-exports.schemas = object.schemas; // OMA LWM2M Objects
+var shared = require('./services/shared');
+exports.Schema = shared.Schema; // constructor
+exports.schemas = shared.schemas; // OMA LWM2M Objects
