+++ lib/services/shared/device.js
