+++ lib/services/shared/security.js
