+++ lib/services/shared/senml.js
